+ Changes to support the ppc sles 9 port of Condor.

+++ src/condor_sysapi/load_avg_t.c
@@ -30,10 +30,10 @@
 #define ITERATIONS_PER_SECOND 10000000
 
 int load_avg_test(int trials, int interval, int num_children, double warn_ok_ratio) {
-	int		foo = 0;
-	int		foo2 = 0;
-	int		bar = 0;
-	int		bar2 = 0;
+	float		foo = 0;
+	float		foo2 = 0;
+	float		bar = 0;
+	float		bar2 = 0;
 	int		i, j, k;
 	int		unslept;
 	int		pid;
